backend/requirements.txt
torch
transformers
pillow
pytesseract
qdrant-client
langchain-community
pymc
fastapi
uvicorn
redis
newsapi-python
alpha-vantage
python-dotenv